+++ src/users.rs
@@ -1,3 +1,4 @@
+use crate::errors::Error;
 use serde::Deserialize;
 
 #[derive(Deserialize, Debug, PartialEq, Eq)]
@@ -8,4 +9,9 @@ pub struct User {
 #[derive(Deserialize, Debug, PartialEq, Eq)]
 pub struct TzInfo {
     pub timezone: String,
+}
+
+pub fn json_to_user(json: String) -> Result<User, Error> {
+    let user: User = serde_json::from_str(&json)?;
+    Ok(user)
 }