+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased (on main branch only)
 
 - Add `section create`
+- Modify some functions in `todoist` module to return structs instead of strings
 
 
 ## 2025-04-30 v0.7.1